Check Balance for Outflows.

+++ Services/IDebtService.cs
@@ -13,6 +13,6 @@ namespace marpha.Services
         Task SaveDebtsAsync(List<Debt> debts);
         Task<List<Debt>> GetAllDebtsAsync();
         Task<int> GetDebtsCountAsync();
-        Task<int> GetTotalDebtAsync();
+        Task<double> GetTotalDebtAsync();
     }
 }
